Separated entities into different files, added exceptions, added hierarchy of engine and wheel, modified calculation of current speed.
