+++ Auto.java
@@ -1,0 +1,12 @@
+package JavaAutomationLab;
+
+public class Auto extends Vehicle {
+    public Auto(String name, Float mass, Engine engine, Wheel wheel) {
+        super(
+                name,
+                mass,
+                engine,
+                wheel
+        );
+    }
+}
