+++ CityRoad.java
@@ -1,0 +1,12 @@
+package JavaAutomationLab;
+
+public class CityRoad extends RouteSection {
+    public CityRoad(PointLocation pointStart, PointLocation pointEnd) {
+        super(
+                pointStart,
+                pointEnd,
+                "CityRoad",
+                0.7F
+        );
+    }
+}
