+++ CountryRoad.java
@@ -1,0 +1,12 @@
+package JavaAutomationLab;
+
+public class CountryRoad extends RouteSection {
+    public CountryRoad(PointLocation pointStart, PointLocation pointEnd) {
+        super(
+                pointStart,
+                pointEnd,
+                "CountryRoad",
+                0.5F
+        );
+    }
+}
