+++ Engine.java
@@ -1,0 +1,13 @@
+package JavaAutomationLab;
+
+public class Engine {
+    private float enginePower;
+
+    public Engine(float enginePower) {
+        this.enginePower = enginePower;
+    }
+
+    public float getEnginePower() {
+        return enginePower;
+    }
+}
