+++ Motorbike.java
@@ -1,0 +1,12 @@
+package JavaAutomationLab;
+
+public class Motorbike extends Vehicle {
+    public Motorbike(String name, Float mass, Engine engine, Wheel wheel) {
+        super(
+                name,
+                mass,
+                engine,
+                wheel
+        );
+    }
+}
