+++ PointLocation.java
@@ -1,0 +1,6 @@
+package JavaAutomationLab;
+
+public class PointLocation {
+    private int x = 0;
+    private int y = 0;
+}
