+++ Truck.java
@@ -1,0 +1,12 @@
+package JavaAutomationLab;
+
+public class Truck extends Vehicle {
+    public Truck(String name, Float mass, Engine engine, Wheel wheel) {
+        super(
+                name,
+                mass,
+                engine,
+                wheel
+        );
+    }
+}
