+++ V4Engine.java
@@ -1,0 +1,10 @@
+package JavaAutomationLab;
+
+public class V4Engine extends Engine {
+
+    private static final int numberOfCylinders = 4;
+
+    public V4Engine() {
+        super(1.4f);
+    }
+}
