+++ V6Engine.java
@@ -1,0 +1,10 @@
+package JavaAutomationLab;
+
+public class V6Engine extends Engine {
+
+    private static final int numberOfCylinders = 6;
+
+    public V6Engine() {
+        super(1.6f);
+    }
+}
