+++ V8Engine.java
@@ -1,0 +1,10 @@
+package JavaAutomationLab;
+
+public class V8Engine extends Engine {
+
+    private static final int numberOfCylinders = 8;
+
+    public V8Engine() {
+        super(1.8f);
+    }
+}
