+++ Wheel.java
@@ -1,0 +1,6 @@
+package JavaAutomationLab;
+
+public abstract class Wheel {
+
+    public abstract float getFrictionCoef(String materialName) throws UnknownRoadMaterialException;
+}
